+++ src/events/onceReady.js
@@ -10,8 +10,8 @@ module.exports = class extends Event {
 	}
 
 	async run() {
-		if (this.client.user.bot) await this.client.fetchApplication();
-		if (!this.client.options.ownerID) this.client.options.ownerID = this.client.user.bot ? this.client.application.owner.id : this.client.user.id;
+		await this.client.fetchApplication();
+		if (!this.client.options.ownerID) this.client.options.ownerID = this.client.application.owner.id;
 
 		this.client.configs = this.client.gateways.clientStorage.get(this.client.user.id, true);
 		await this.client.gateways.sync();