+++ src/lib/util/util.js
@@ -52,11 +52,7 @@ class Util {
 	 * @param {KlasaClient} client The Klasa client
 	 */
 	static initClean(client) {
-		const patterns = [];
-		if (client.token) patterns.push(Util.regExpEsc(client.token));
-		if (client.user.email) patterns.push(Util.regExpEsc(client.user.email));
-		if (client.password) patterns.push(Util.regExpEsc(client.password));
-		sensitivePattern = new RegExp(patterns.join('|'), 'gi');
+		sensitivePattern = new RegExp(Util.regExpEsc(client.token), 'gi');
 	}
 
 	/**